# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "get": {
        "operationId": "ProductsController_list",
        "summary": "List ACTIVE products with filter/search/sort/pagination.",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Field to sort by (allow-listed per resource).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Full-text-ish search over name/description.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category slug filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price in cents.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price in cents.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (public)"
        ]
      }
    },
    "/api/products/{slug}": {
      "get": {
        "operationId": "ProductsController_getBySlug",
        "summary": "Get a single ACTIVE product by slug.",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (public)"
        ]
      }
    },
    "/api/products/{slug}/related": {
      "get": {
        "operationId": "ProductsController_related",
        "summary": "Related products in the same categories.",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (public)"
        ]
      }
    },
...
    "/api/admin/products/{id}": {
      "get": {
        "operationId": "AdminProductsController_get",
        "summary": "Get a product by id (admin).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AdminProductsController_update",
        "summary": "Update a product (reconciles variants + categories).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_remove",
        "summary": "Delete a product.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nexlor Commerce API",
    "description": "Storefront + Admin API for the Nexlor commerce template.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VariantInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "priceOverride": {
            "type": "object",
            "description": "Override price in cents; null = use basePrice."
          }
        }
      },
      "ImageInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "alt": {
            "type": "string"
          }
        }
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "URL slug; auto-generated from name if omitted."
          },
          "basePrice": {
            "type": "number",
            "description": "Base price in cents."
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "categoryIds": {
            "description": "Category ids.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantInputDto"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageInputDto"
            }
          }
        }
      },
...
    }
  }
}